CNN using TensorFlow for image classification
Here we set up our network.  The best results I was able to obtain were from the architecture:
 56x56x3 input --> 5x5x32 convolution layer --> 2x2 maxpool layer -> 1024 fully connected layer (relu) -> 3 logits with softmax
 Obviously many other setups are possible, and this code can be changed with a couple keystrokes to add
another fully connected layer with dropout or another conv-pool set of layers.
For this problem, due to data scarcity, these additional layers did not contribute positively to the final result
 If training for a long time, however, using the dropout layer was seen to combat overfitting

The best results obtained had an accuracy of 76%.  Comparable test scores were obtained with and without dropout, although in the latter case the training scores approached 100%, which is obviously an indication of overfitting.

Nevertheless, given the relatively small size of the dataset, and the inconsistency of the images, I am quite proud of the 76% classification accuracy.